Move extract_bounds into a common netcdf utility module
